+++ ktrain/tests/test_lda.py
@@ -42,10 +42,10 @@ class TestLDA(TestCase):
         # document similarity
         tech_topics = [51, 85, 94, 22]
         tech_probs = tm.get_doctopics(topic_ids=tech_topics)
-        doc_ids = [doc[1] for doc in tm.get_docs(topic_ids=tech_topics)]
+        doc_ids = [doc['doc_id'] for doc in tm.get_docs(topic_ids=tech_topics)]
         tm.train_scorer(topic_ids=tech_topics)
         other_topics = [i for i in range(tm.n_topics) if i not in tech_topics]
-        other_texts = [d[0] for d in tm.get_docs(topic_ids=other_topics)]
+        other_texts = [d['text'] for d in tm.get_docs(topic_ids=other_topics)]
         other_scores = tm.score(other_texts)
         # display results in Pandas dataframe
         other_preds = [int(score > 0) for score in other_scores]
@@ -56,7 +56,7 @@ class TestLDA(TestCase):
         # recommender
         tm.train_recommender()
         results = tm.recommend(text=rawtext, n=1)
-        self.assertTrue(results[0][0].startswith('Archive-name'))
+        self.assertTrue(results[0]['text'].startswith('Archive-name'))
 
 
 if __name__ == "__main__":