+++ ktrain/text/preprocessor.py
@@ -298,7 +298,8 @@ def hf_convert_examples(texts, y=None, tokenizer=None,
                         pad_on_left=False,
                         pad_token=0,
                         pad_token_segment_id=0,
-                        mask_padding_with_zero=True):
+                        mask_padding_with_zero=True,
+                        verbose=1):
     """
     Loads a data file into a list of ``InputFeatures``
     Args:
@@ -312,6 +313,7 @@ def hf_convert_examples(texts, y=None, tokenizer=None,
         mask_padding_with_zero: If set to ``True``, the attention mask will be filled by ``1`` for actual values
             and by ``0`` for padded values. If set to ``False``, inverts it (``1`` for padded values, ``0`` for
             actual values)
+        verbose(bool): verbosity
     Returns:
         If the ``examples`` input is a ``tf.data.Dataset``, will return a ``tf.data.Dataset``
         containing the task-specific features. If the input is a list of ``InputExamples``, will return
@@ -320,11 +322,17 @@ def hf_convert_examples(texts, y=None, tokenizer=None,
 
     is_array, is_pair = detect_text_format(texts)
     data = []
-    mb = master_bar(range(1))
     features_list = []
     labels = []
+    if verbose:
+        mb = master_bar(range(1))
+        pb = progress_bar(texts, parent=mb)
+    else:
+        mb = range(1)
+        pb = texts
     for i in mb:
-        for (idx, text) in enumerate(progress_bar(texts, parent=mb)):
+        #for (idx, text) in enumerate(progress_bar(texts, parent=mb)):
+        for (idx, text) in enumerate(pb):
             if is_pair:
                 text_a = text[0]
                 text_b = text[1]
@@ -895,7 +903,8 @@ class TransformersPreprocessor(TextPreprocessor):
         dataset = hf_convert_examples(texts, y=y, tokenizer=tok, max_length=self.maxlen,
                                       pad_on_left=bool(self.name in ['xlnet']),
                                       pad_token=tok.convert_tokens_to_ids([tok.pad_token][0]),
-                                      pad_token_segment_id=4 if self.name in ['xlnet'] else 0)
+                                      pad_token_segment_id=4 if self.name in ['xlnet'] else 0,
+                                      verbose=verbose)
         self.set_multilabel(dataset, mode, verbose=verbose)
         if mode == 'train':  self.preprocess_train_called = True
         return dataset