+++ CHANGELOG.md
@@ -6,6 +6,22 @@ Most recent releases are shown at the top. Each release shows:
 - **Changed**: Additional parameters, changes to inputs or outputs, etc
 - **Fixed**: Bug fixes that don't change documented behaviour
 
+## 0.23.0 (TBD)
+
+### New:
+- Support for upcoming release of TensorFlow 2.4 such as removal of references to obsolete `multi_gpu_model`
+
+### Changed
+- **[breaking change]** `TopicModel.get_docs` now returns a list of dicts instead of a list of tuples.  Each dict has keys: `text`, `doc_id`, `topic_proba`, `topic_id`.
+- added `TopicModel.get_document_topic_distribution` 
+- added `TopicModel.get_sorted_docs` method to return all documents sorted by relevance to a given `topic_id`
+
+
+### Fixed:
+- Changed version check warning in `lr_find` to a raised Exception to avoid confusion when warnings from **ktrain** are suppressed
+- Pass `verbose` parameter to `hf_convert_examples`
+
+
 ## 0.22.4 (2020-10-12)
 
 ### New: